wahoo: Create more inodes for flexibility
We routinely run out of inodes on /system if we limit it so tightly
(~4096) due to ever-increasing amounts of files as more packages and
libraries are added in newer versions of Android and downstream
extensions.

Instead of constantly having to increase the inode count, just set it to
8192 using the existing PRODUCT_INCREASE_INODE_COUNT flag to fix the
problems permanently. 8192 inodes should be more than enough for most
purposes, and it's still low enough that the increased space usage is
negligible.

Change-Id: I59d866606fbdd06f0959a8c08db07a748532bdcb